+++ sys/hp300/dev/ppi.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)ppi.c	8.1 (Berkeley) 06/16/93
+ *	@(#)ppi.c	8.2 (Berkeley) 01/09/95
  */
 
 /*
@@ -325,7 +325,7 @@ again:
 
 ppiioctl(dev, cmd, data, flag)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 {
